Adding custom exception and global exception hamdling support

+++ src/main/java/com/example/Spring/Boot/Blog/service/impl/PostServiceImpl.java
@@ -2,16 +2,17 @@ package com.example.Spring.Boot.Blog.service.impl;
 
 import com.example.Spring.Boot.Blog.dto.PostDto;
 import com.example.Spring.Boot.Blog.dto.PostResponse;
+import com.example.Spring.Boot.Blog.exception.ResourceNotFoundException;
 import com.example.Spring.Boot.Blog.model.Post;
 import com.example.Spring.Boot.Blog.repository.PostRepository;
 import com.example.Spring.Boot.Blog.service.PostService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -19,10 +20,12 @@ import java.util.stream.Collectors;
 @Service
 public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
+    private final ModelMapper mapper;
 
     @Autowired
-    public PostServiceImpl(PostRepository postRepository) {
+    public PostServiceImpl(PostRepository postRepository, ModelMapper mapper) {
         this.postRepository = postRepository;
+        this.mapper = mapper;
     }
 
     @Override
@@ -38,19 +41,19 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public PostResponse getAllPost(int pageNo, int pageSize,String sortBy,String sortDir) {
+    public PostResponse getAllPost(int pageNo, int pageSize, String sortBy, String sortDir) {
         //Getting the sort direction
-        Sort sort=sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending():
+        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() :
                 Sort.by(sortBy).descending();
         //create Pageable Instances
-        Pageable pageable = PageRequest.of(pageNo, pageSize,sort);
+        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
         Page<Post> posts = postRepository.findAll(pageable);
 
         //Get all pages
         List<Post> allPosts = posts.getContent();
-        List<PostDto> content= allPosts.stream().map(t -> mapToDto(t)).collect(Collectors.toList());
+        List<PostDto> content = allPosts.stream().map(t -> mapToDto(t)).collect(Collectors.toList());
 
-        PostResponse postResponse=new PostResponse();
+        PostResponse postResponse = new PostResponse();
         postResponse.setContent(content);
         postResponse.setPageNo(posts.getNumber());
         postResponse.setPageSize(posts.getSize());
@@ -64,51 +67,39 @@ public class PostServiceImpl implements PostService {
     @Override
     public PostDto getPostById(long id) {
         Optional<Post> post = postRepository.findById(id);
-        if (post.isPresent()) {
-            return mapToDto(post.get());
-        }
-        return null;
+        if (!post.isPresent())
+            throw new ResourceNotFoundException("post","id",id);
+        return mapToDto(post.get());
     }
 
     @Override
     public PostDto updatePostById(PostDto postDto, long id) {
         Optional<Post> post = postRepository.findById(id);
-        if (post.isPresent()) {
-            Post fetchedPost = post.get();
-            Post updatedPost = mapToEntity(postDto);
-            fetchedPost.setTitle(updatedPost.getTitle());
-            fetchedPost.setDescription(updatedPost.getDescription());
-            fetchedPost.setContent(updatedPost.getContent());
-            Post newlyUpdatedAndSavedPost = postRepository.save(fetchedPost);
-            return mapToDto(newlyUpdatedAndSavedPost);
-        }
-        return null;
+        if (!post.isPresent())
+              throw new ResourceNotFoundException("post","id",id);
+        Post fetchedPost = post.get();
+        Post updatedPost = mapToEntity(postDto);
+        fetchedPost.setTitle(updatedPost.getTitle());
+        fetchedPost.setDescription(updatedPost.getDescription());
+        fetchedPost.setContent(updatedPost.getContent());
+        Post newlyUpdatedAndSavedPost = postRepository.save(fetchedPost);
+        return mapToDto(newlyUpdatedAndSavedPost);
     }
 
     @Override
-    public boolean deletePostById(long id) {
+    public void deletePostById(long id) {
         Optional<Post> post = postRepository.findById(id);
-        if (post.isPresent()) {
-            postRepository.deleteById(id);
-            return true;
-        }
-        return false;
+        if(post.isPresent()) postRepository.deleteById(id);
+        else throw  new ResourceNotFoundException("post","id",id);
     }
 
     private PostDto mapToDto(Post post) {
-        PostDto postDto = new PostDto();
-        postDto.setId(post.getId());
-        postDto.setTitle(post.getTitle());
-        postDto.setContent(post.getContent());
-        postDto.setDescription(post.getDescription());
+        PostDto postDto = mapper.map(post, PostDto.class);
         return postDto;
     }
 
     private Post mapToEntity(PostDto postDto) {
-        Post post = new Post();
-        post.setTitle(postDto.getTitle());
-        post.setContent(postDto.getContent());
-        post.setDescription(postDto.getDescription());
+        Post post = mapper.map(postDto, Post.class);
         return post;
     }
 }
